tools/arxiv-tracker/requirements.txt
# FormalMath arXiv前沿跟踪系统依赖

# 核心依赖
requests>=2.28.0          # HTTP请求
pyyaml>=6.0               # YAML配置解析

# 可选依赖（用于高级功能）
# numpy>=1.21.0           # 数值计算
# pandas>=1.3.0           # 数据处理
# scikit-learn>=1.0.0     # 机器学习（用于高级相关性评分）

# 调度（用于定时任务）
# schedule>=1.1.0         # 简单调度
# apscheduler>=3.9.0      # 高级调度

# 开发依赖
# pytest>=7.0.0           # 测试框架
# black>=22.0.0           # 代码格式化
# flake8>=4.0.0           # 代码检查
